dont throw an exception if trying to remove a previous report that do…

+++ service/src/main/java/eu/europa/ec/fisheries/uvms/movementrules/service/dao/RulesDao.java
@@ -220,7 +220,9 @@ public class RulesDao {
     }
 
     public void deletePreviousReport(PreviousReport report) {
-        em.remove(report);
+        if(report != null) {
+            em.remove(report);
+        }
     }
 
     public long getNumberOfTicketsForRule(String ruleGuid){
